backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
argon2-cffi
fastapi-login
pyjwt
python-jose
email-validator
pydantic[email]
redis
python-multipart
itsdangerous
langchain-community
langchain
pinecone-client
PyPDF2
openai
numpy
azure-storage-blob
azure-cosmos
azure-identity
azure-keyvault-secrets
langchain_openai
langchain_pinecone